arm64: dts: ls1028a: sl28: use ocelot-8021q tagging by default
Enable the ocelot-8021q tagger by default which supports ethernet flow
control.

The new default is set in the common board dtsi. The actual switch
node is enabled on a per board variant basis. Because of this we
set the new tagger default for both internal ports and a particular
variant is free to choose among the two port.

Signed-off-by: Michael Walle <michael@walle.cc>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: be0b178c50c37a666d54f435da71cf9f008362a0 ]