dt-bindings: display: mediatek: Update disp_aal binding for MT8183
The driver data of MT8183 and MT8173 are different.

For MT8173, the gamma module is inside disp_aal. When we need to adjust
gamma value, we need to use "has_gamma" to control gamma function
inside disp_aal to adjust the gamma value.

For successors like MT8183, disp_gamma is separated from disp_aal. We
just need to control disp_gamma directly and don't need to control gamma
function inside disp_aal.

With this modification, the driver doesn't require any functional changes.
We only update the dt-binding and DTS node to make it clear.

Link: https://patchwork.kernel.org/project/linux-mediatek/patch/20220411035843.19847-2-rex-bc.chen@mediatek.com/
Signed-off-by: Rex-BC Chen <rex-bc.chen@mediatek.com>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Chun-Kuang Hu <chunkuang.hu@kernel.org>

[ upstream commit: 402fc936d5a9328779f175893b30ae165d45a11e ]