dt-bindings: Drop undocumented i.MX iomuxc-gpr bindings in examples
The i.MX iomuxc-gpr bindings are undocumented and a mess. Drop their use
from the examples.

The problem with the binding beyond the just random variations is that
the iomuxc-gpr is not a separate block, but registers within the iomuxc
block containing random leftover controls. As a separate DT node, it
creates nodes with overlapping memory addresses.

Signed-off-by: Rob Herring <robh@kernel.org>
Acked-by: Peter Rosin <peda@axentia.se>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Link: https://lore.kernel.org/r/20220422192121.2592030-1-robh@kernel.org

[ upstream commit: 3474b838f4204c21d108183d9268611d961a428f ]