dt-bindings: vexpress-sysreg: Allow for no child nodes
It is valid for arm,vexpress-sysreg to have no child nodes when either
the deprecated binding without child nodes is used or a platform
has no GPIOs. If there are no child nodes, then "#address-cells" and
"#size-cells" should not be required. However, "reg" should be present
in all cases, so let's add it now.

Cc: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20220613171809.3928949-2-robh@kernel.org

[ upstream commit: cfda05c1ddb8b375562b54dc86276cdc42d48297 ]