dt-bindings: arm: renesas: Document R-Car H3Ne-1.7G SoC and boards
Document the compatible values for the R-Car H3Ne-1.7G (R8A779MB)
SoC.  This is a different grading of the R-Car H3-N (R8A77951) SoC.

All R-Car H3Ne-1.7G on-SoC devices are identical to the corresponding
devices on the R-Car H3-N SoC, and thus just use the compatible values
for the latter.  The root compatible property does gain an additional
value, to sort out integration issues if they ever arise.

Document the use of this SoC on the Salvator-XS and ULCB (with and
without Kingfisher) development boards.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/f816f58d1c3fbd0605589e7648a43d78a4e1e694.1656072871.git.geert+renesas@glider.be

[ upstream commit: f5f6306965aaa9a9db8ebb02d725759eeee24b24 ]