dt-bindings: input: touchscreen: add compatible string for Goodix GT1158
Goodix GT1158 is a touchscreen chip from Goodix used in the PinePhone
and PinePhone Pro. Patches to correct these devices dts files will be
sent in a later patch series.

This driver was modified to support the GT1158 in the patch linked
below. Add its compatible string to the device tree binding.

Suggested-by: Ondrej Jirman <megi@xff.cz>
Signed-off-by: Jarrah Gosbell <kernel@undef.tools>
Link: https://lore.kernel.org/r/20220809101633.352315-1-kernel@undef.tools
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

[ upstream commit: b142af334de60f669f36461036360867babcc912 ]